+++ server/index.js
@@ -1,9 +1,9 @@
 const server = require('./src/server');
 const { conn } = require('./src/db');
 require('dotenv').config();
-const clientLoader = require('./src/controllers/loadClient');
-const professionalLoader = require('./src/controllers/loadProfessional');
-const serviceLoader = require('./src/controllers/loadservice');
+const clientLoader = require('./src/assets/funtions/loadClient');
+const professionalLoader = require('./src/assets/funtions/loadProfessional');
+const serviceLoader = require('./src/assets/funtions/loadservice');
 //const PORT = 3001;
 const PORT = process.env.PORT
