+++ server/src/assets/funtions/loadClient.js
