+++ server/src/controllers/appointmentControllers.js
@@ -1,0 +1,5 @@
+const {Appointment}= require ('../db')
+
+//controllers de appointments
+
+module.exports={}
