+++ server/src/controllers/paymentControllers.js
@@ -1,0 +1,5 @@
+const {Payment}= require ('../db')
+
+//controllers de payment
+
+module.exports={}
