+++ server/src/controllers/professionalControllers.js
@@ -1,0 +1,5 @@
+const {Professional}= require ('../db')
+
+//controllers de professional
+
+module.exports={}
