+++ server/src/handlers/appointmentHandler.js
@@ -1,0 +1,6 @@
+const {} = require('../controllers/appointmentControllers');
+
+//handlers para appointment
+
+
+module.exports= {}
