+++ server/src/handlers/paymentHandler.js
@@ -1,0 +1,6 @@
+const {} = require('../controllers/paymentControllers');
+
+//handlers para payment
+
+
+module.exports= {}
