+++ server/src/handlers/professionalHandler.js
@@ -1,0 +1,5 @@
+const {} = require('../controllers/professionalControllers');
+
+//handlers para professional
+
+module.exports= {}
