+++ server/src/routes/appointmentRouter.js
@@ -1,0 +1,9 @@
+const {Router} = require('express');
+const {} = require('../handlers/appointmentHandler');
+const appointmentRouter = Router();
+
+
+// appointmentRouter.get('/',handler); 
+
+
+module.exports = appointmentRouter;
