+++ server/src/routes/paymentRouter.js
@@ -1,0 +1,10 @@
+const {Router} = require('express');
+const { } = require('../handlers/paymentHandler'); 
+const paymentRouter = Router();
+
+
+// paymentRouter.get('/', handler);
+
+
+
+module.exports = paymentRouter;
