+++ server/src/routes/professionalRouter.js
@@ -1,0 +1,10 @@
+const {Router} = require('express');
+const { } = require('../handlers/professionalHandler'); 
+const professionalRouter = Router();
+
+
+// professionalRouter.get('/', handler);
+
+
+
+module.exports = professionalRouter;
