+++ server/src/routes/serviceRouter.js
@@ -1,0 +1,10 @@
+const {Router} = require('express');
+const {getServices, postService } = require('../handlers/serviceHandler'); 
+const serviceRouter = Router();
+
+
+serviceRouter.get('/',getServices);
+serviceRouter.post('/',postService); 
+
+
+module.exports = serviceRouter;
